refactor: hidden 속성을 제어하는 함수 util 역할로 분리
