+++ src/js/utils/attribute.js
@@ -1,0 +1,7 @@
+export const hideElement = (element) => {
+  return element.setAttribute('hidden', true);
+};
+
+export const showElement = (element) => {
+  return element.removeAttribute('hidden');
+};
